include hardcoded security

+++ src/main/java/com/nasim/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.nasim.model.User;
 
-public interface UserRepository extends PagingAndSortingRepository<User, Integer> {
+public interface UserRepository extends PagingAndSortingRepository<User, Long> {
 	Optional<User> findByUsername(String username);
 
 	Boolean existsByUsername(String username);
